% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{logLik_fft}
\alias{logLik_fft}
\alias{logLik_fft2}
\title{Compute Log-likelihood via KDE-FFT}
\usage{
logLik_fft(y, yhat, h = NA, m = 0.8, p = 10, nThreads = 2)

logLik_fft2(y, yhat, p = 10, nThreads = 2)
}
\arguments{
\item{y}{a vector storing empirical data (e.g., RTs)}

\item{yhat}{a vector storing simulated data.}

\item{h}{KDE bandwidth}

\item{m}{a multiplier to adjust \code{h} proportationally. Default is 0.8. 
Set \code{m=1}, if one wish not adjust bandwidth.}

\item{p}{a precision parameter defines the number of grid as power of 2.
Default value is 10 (i.e., \code{ngrid=2^10}). This is for \code{logLik_fft}.}

\item{nThreads}{number of threads per block. Default is 2}
}
\value{
Log-likelihood.
}
\description{
This function uses GPU to calculate log-likelihood via KDE-FFT 
algorithm.
}
\details{
'\code{logLik_fft1} returns summed log-likelihood.  \code{logLik_fft2} returns
a matrix storing individual data point (column 1) and log-likelihood for 
these data points (column 2).
}
\examples{
## ----- Use piecewise LBA data  ----- ## 
## Use piecewise LBA data as an example ----
data(lba); bandwidth <- .02;
tmp1 <- cpda::logLik_fft(plba$DT1, plba$eDT1, bandwidth)
tmp2 <- gpda::logLik_fft(plba$DT1, plba$eDT1, bandwidth)
tmp1
tmp2

tmp3 <- cpda::logLik_pw(plba$DT2, plba$eDT2, bandwidth)
tmp4 <- gpda::logLik_pw(plba$DT2, plba$eDT2, bandwidth)
sum(tmp3) 
sum(tmp4)

## Set ngrid to 2^20
tmp5 <- gpda::logLik_fft(plba$DT1, plba$eDT1, bandwidth, p=20)

## ----- Use MATLAB data ----- ## 
tmp1 <- read.table("tests/Datavec.txt") ## From matlab
y <- as.double(sort(tmp1$V1))
tmp2 <- read.table("tests/sampvec.txt") ## From matlab
yhat <- as.double(sort(tmp2$V1))
## or use R to simulate; -1435.704
yhat <- as.double(sort(rnorm(ns, 5, 1)))
ns <- as.integer(length(yhat))
ny <- as.integer(length(y))

## Calculate pointwise likelihood  
LL3 <- gpda::logLik_fft(y, yhat, pointwise=TRUE); 
LL4 <- gpda::logLik_fft2(plba$DT1, plba$eDT1, bandwidth)
sum(LL4[,2])  ## Column 1 store data; column 2 store log-likelihoods

}
\references{
Holmes, W. (2015). A practical guide to the Probability Density
Approximation (PDA) with improved implementation and error characterization.
\emph{Journal of Mathematical Psychology}, \bold{68-69}, 13--24,
doi: http://dx.doi.org/10.1016/j.jmp.2015.08.006.
}
