% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{n1PDF_plba1}
\alias{n1PDF_plba1}
\alias{n1PDF_plba2}
\alias{n1PDF_plba3}
\title{Approximate Node 1 Likelihood of pLBA Model}
\usage{
n1PDF_plba1(x, nsim = 1024, b = 2.7, A = 1.5, mean_v = c(3.3, 2.2),
  mean_w = c(1.5, 1.2), sd_v = c(1, 1), rD = 0.3, swt = 0.5,
  t0 = 0.08, nthread = 32, debug = FALSE)

n1PDF_plba2(x, nsim = 1024, b = c(2.7, 2.7), A = c(1.5, 1.5),
  mean_v = c(3.3, 2.2), mean_w = c(1.5, 3.7), sd_v = c(1, 1),
  sd_w = c(1.5, 1.2), rD = 0.3, swt = 0.5, t0 = 0.08, nthread = 32,
  debug = FALSE)

n1PDF_plba3(x, nsim = 1024, B = c(1.2, 1.2), A = c(1.5, 1.5), C = c(0.3,
  0.3), mean_v = c(3.3, 2.2), mean_w = c(1.5, 3.7), sd_v = c(1, 1),
  sd_w = c(1, 1), rD = 0.1, tD = 0.1, swt = 0.5, t0 = 0.08,
  nthread = 32, debug = FALSE)
}
\arguments{
\item{x}{a data vector for estimating likelihood.}

\item{nsim}{number of simulations. This must be a power of two.}

\item{b}{threshold. Must be a scalar for plba1. Must be a two-element vector
for plba2.}

\item{A}{starting point upper bound. Must be a scalar for plba1. Must be a 
two-element vector for plba2 and plba3.}

\item{mean_v}{mean drift rate stage 1. This must be a two-element vector.}

\item{mean_w}{mean drift rate stage 2. This must be a two-element vector.}

\item{sd_v}{standard deviation of drift rate stage 1. This must be a 
two-element vector.}

\item{rD}{an internal psychological delay time for drift rate.}

\item{swt}{an external switch time when task information changes.}

\item{t0}{non-decision time.}

\item{nthread}{numbers of launched GPU threads. Default is a wrap.}

\item{debug}{a debugging switch.}

\item{B}{travelling distance stage 1. The distance between starting point (
drawn randomly from an uniform distribution) to the threshold.  This applies
for plba3 only. Please note B differs from b. Must be a two-element vector
for plba2.}

\item{C}{travelling distance stage 2. The distance between updated threshold 
and original threshold This applies for plba3 only. Must be a two-element 
vector. Note this is a uppercase \bold{C}.}

\item{tD}{an internal psychological delay time for threshold. This applies
for plba3 only.}
}
\value{
a likelihood vector.
}
\description{
This is the approximated density function for 2-accumualtor piecewise LBA 
model, sampling drift rates from truncated normal distributions. The 
function uses probability density approximation to estimate likelihood.
}
\examples{
n <- 2^20
x <- seq(0, 3, length.out = 1e3);

## plba1
den1 <- gpda::n1PDF_plba1(x, nsim=n, b=2.7, A=1.5, mean_v=c(3.3, 2.2), 
                          mean_w=c(1.5, 1.2), sd_v=c(1, 1), rD=.3, swt=.5,
                          t0=.08)
plot(x, den1, type="l")

## plba2
den1 <- gpda::n1PDF_plba2(data, nsim=n, b=c(2.7, 2.7), A=c(1.5,1.5),
mean_v=c(3.3, 2.2), mean_w=c(1.5, 1.2),
sd_v=c(1, 1), rD=.3, swt=.5, t0=.08)
plot(x, dat1, type="l")

## plba3
pvec1 <- c(A1 = 1.51, A2 = 1.51, B1 = 1.2, B2 = 1.2,   C1 = .3, C2 = .3,
  v1 = 3.32, v2 = 2.24, w1 = 1.51, w2 = 3.69, sv1 = 1, sv2 = 1,
  sw1 = 1, sw2 = 1, rD = 0.1, tD = .1, swt = 0.5, t0 = 0.08)

pvec2 <- c(A1 = 1.51, A2 = 1.51, B1 = 1.2, B2 = 1.2, C1 = .3, C2 = .3,
  v1 = 3.32, v2 = 2.24, w1 = 1.51, w2 = 3.69, sv1 = 1, sv2 = 1,
  sw1 = 1, sw2 = 1, rD = 0.1, tD = .15, swt = 0.5, t0 = 0.08)
pvec3 <- c(A1 = 1.51, A2 = 1.51, B1 = 1.2, B2 = 1.2, C1 = .3, C2 = .3,
  v1 = 3.32, v2 = 2.24, w1 = 1.51, w2 = 3.69, sv1 = 1, sv2 = 1,
  sw1 = 1, sw2 = 1, rD = 0.15, tD = .1, swt = 0.5, t0 = 0.08)

den3 <- gpda::n1PDF_plba3(data, nsim = n, B = pvec1[3:4], A =pvec1[1:2],
  C = pvec1[5:6], mean_v = pvec1[7:8],
  mean_w = pvec1[9:10], sd_v = pvec1[11:12],
  sd_w = pvec1[13:14], rD = pvec1[15], tD = pvec1[16],
  swt = pvec1[17], t0 = pvec1[18])

plot(x, tmp, type="l")

}
\references{
Holmes, W., Trueblood, J. S., & Heathcote, A. (2016). A new 
framework for modeling decisions about changing information: The Piecewise 
Linear Ballistic Accumulator model \emph{Cognitive Psychology}, \bold{85},
1--29, \cr doi: \url{http://dx.doi.org/10.1016/j.cogpsych.2015.11.002}.
}
