% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rplba0}
\alias{rplba0}
\alias{rplba1}
\alias{rplba2}
\alias{rplba3}
\title{The Random Number Generator of the pLBA Model}
\usage{
rplba0(n, A = 1.5, b = 2.7, t0 = 0.5, mean_v = c(3.3, 2.2),
  mean_w = c(1.5, 1.2), sd_v = c(1, 1), rD = 0.3, swt = 0.5,
  nthread = 32, gpuid = 0)

rplba1(n, A = 1.5, b = 2.7, t0 = 0.5, mean_v = c(3.3, 2.2),
  mean_w = c(1.5, 1.2), sd_v = c(1, 1), rD = 0.3, swt = 0.5,
  nthread = 32, gpuid = 0)

rplba2(n, A = c(1.5, 1.5), b = c(2.7, 2.7), t0 = 0.08, mean_v = c(3.3,
  2.2), mean_w = c(1.5, 3.7), sd_v = c(1, 1), sd_w = c(1, 1), rD = 0.3,
  swt = 0.5, nthread = 32, gpuid = 0)

rplba3(n, A = c(1.5, 1.5), B = c(1.2, 1.2), C = c(0.3, 0.3),
  mean_v = c(3.3, 2.2), mean_w = c(1.5, 3.7), sd_v = c(1, 1),
  sd_w = c(1, 1), rD = 0.3, tD = 0.3, swt = 0.5, t0 = 0.08,
  nthread = 32, gpuid = 0)
}
\arguments{
\item{n}{number of simulations. This must be a power of two.}

\item{A}{starting point upper bound. Must be a scalar for plba1. Must be a 
two-element vector for plba2 and plba3.}

\item{b}{threshold. Must be a scalar for plba1. Must be a two-element vector
for plba2.}

\item{t0}{non-decision time.}

\item{mean_v}{mean drift rate stage 1. This must be a two-element vector.}

\item{mean_w}{mean drift rate stage 2. This must be a two-element vector.}

\item{sd_v}{standard deviation of drift rate stage 1. This must be a 
two-element vector.}

\item{rD}{an internal psychological delay time for drift rate.}

\item{swt}{an external switch time when task information changes.}

\item{nthread}{numbers of launched GPU threads. Default is a wrap.}

\item{gpuid}{select which GPU to conduct model simulation, if running on 
multiple GPU machines.}

\item{sd_w}{standard deviation of drift rate stage 2. This must be a 
two-element vector.}

\item{B}{travelling distance stage 1. The distance between starting point (
drawn randomly from an uniform distribution) to the threshold.  This applies
for plba3 only. Please note B differs from b. Must be a two-element vector.}

\item{C}{travelling distance stage 2. The distance between updated threshold 
and original threshold This applies for plba3 only. Must be a two-element 
vector. Note this is uppercase.}

\item{tD}{an internal psychological delay time for threshold. This applies
for plba3 only.}
}
\value{
a 2-column data frame [RT R].
}
\description{
This function generates two-accumulator pLBA random numbers using GPU.
}
\examples{
n <- 2^20
dat1 <- gpda::rplba1(n)
dat2 <- gpda::rplba2(n)
dat3 <- gpda::rplba3(n)

crt1 <- dat1[dat1$R==1,"RT"]
ert1 <- dat1[dat1$R==2,"RT"]
crt2 <- dat2[dat2$R==1,"RT"]
ert2 <- dat2[dat2$R==2,"RT"]
crt3 <- dat3[dat3$R==1,"RT"]
ert3 <- dat3[dat3$R==2,"RT"]

par(mfrow=c(3,2))
hist(crt1, breaks="fd")
hist(ert1, breaks="fd")
hist(crt2, breaks="fd")
hist(ert2, breaks="fd")
hist(crt3, breaks="fd")
hist(ert3, breaks="fd")
par(mfrow=c(1,1))

## It takes about 10 ms to simulate 2^20 rplba random numbers.
\dontrun{
require(microbenchmark)
res <- microbenchmark(gpda::rplba1(n),
                      gpda::rplba2(n),
                      gpda::rplba3(n), times=10L)
}

## Unit: milliseconds
##             expr      min       lq     mean   median       uq      max 
##  gpda::rplba1(n) 9.046328 10.00658 10.68783 10.72616 11.37464 12.02669 
##  gpda::rplba2(n) 9.349161 11.00560 14.71400 11.19560 13.47059 41.00774 
##  gpda::rplba3(n) 9.870809 10.06734 11.36654 11.19728 11.79733 14.41500 

}
\references{
Holmes, W., Trueblood, J. S., & Heathcote, A. (2016). A new 
framework for modeling decisions about changing information: The Piecewise 
Linear Ballistic Accumulator model \emph{Cognitive Psychology}, \bold{85},
1--29, \cr doi: \url{http://dx.doi.org/10.1016/j.cogpsych.2015.11.002}.
}
