% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rtnorm}
\alias{rtnorm}
\title{Generate Random numbers from a Truncated Normal Distribution}
\usage{
rtnorm(n, mean = 0, sd = 1, lower = -Inf, upper = Inf, dp = FALSE,
  nthread = 32)
}
\arguments{
\item{n}{number of observations. Only accept an integer scalar}

\item{mean}{mean. Only a scalar}

\item{sd}{standard deviations. Only accept a scalar}

\item{lower}{lower bound. Only accept a scalar}

\item{upper}{upper bound. Only accept a scalar}

\item{dp}{whether calculate using double precision. Default is FALSE.}

\item{nthread}{number of threads launched per block.}
}
\value{
a double vector
}
\description{
This function generates random numbers using GPU from a truncated normal 
distribution.
}
\examples{
dat1 <- gpda::rtnorm(n, mean=-1, sd=1.2, lower=0, upper=Inf)

## https://github.com/TasCL/tnorm
## dat2 <- tnorm::rtnorm(n, mean=-1, sd=1.2, lower=0, upper=Inf)
## dat3 <- msm::rtnorm(n, mean=-1, sd=1.2, lower=0, upper=Inf)
den1 <- density(dat1)
## den2 <- density(dat2)
## den3 <- density(dat3)

summary(dat1)
## summary(dat2)
## summary(dat3)

\dontrun{
par(mfrow=c(1,3))
hist(dat2, breaks="fd", freq=FALSE)
lines(den1$x, den1$y,lwd=2) ## gpu
lines(den2$x, den2$y,lwd=2) ## tnorm
lines(den3$x, den3$y,lwd=2) ## msn

hist(dat1, breaks="fd", freq=FALSE)
lines(den1$x, den1$y,lwd=2) ## gpu
lines(den2$x, den2$y,lwd=2) ## tnorm
lines(den3$x, den3$y,lwd=2) ## msn

hist(dat3, breaks="fd", freq=FALSE)
lines(den1$x, den1$y,lwd=2) ## gpu
lines(den2$x, den2$y,lwd=2) ## tnorm
lines(den3$x, den3$y,lwd=2) ## msn
par(mfrow=c(1,1))

## Unit: milliseconds
##            expr       min         lq       mean     median         uq        max
## gpda::rtnorm(n)   1.173537   1.417016   1.978613   1.423757   1.580943   6.976541
## tnorm::rtnorm(n)  7.475374   8.317984   8.544317   8.345958   9.120224  10.220493
## msm::rtnorm(n)   54.597366 109.426265 103.025877 110.050924 119.054471 125.192521
}
}
