% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sum}
\alias{sum}
\alias{min}
\alias{max}
\alias{minmax}
\alias{sqsum}
\alias{sd}
\alias{count}
\title{GPU Math Functions}
\usage{
sum(x, nthread = 32, debug = FALSE)

min(x, nthread = 32, debug = FALSE)

max(x, nthread = 32, debug = FALSE)

minmax(x, nthread = 32, debug = FALSE)

sqsum(x, nthread = 32, debug = FALSE)

sd(x, nthread = 32, debug = FALSE)

count(R, nthread = 32, debug = FALSE)
}
\arguments{
\item{x}{a data vector.}

\item{nthread}{numbers of thread to launch the kernel}

\item{debug}{whether to print debug information}
}
\value{
a scalar or a pair of scalar values.
}
\description{
sum, min, max, minmax and sd are 5 basic mathematical functions to sum a
vector, to find minimal or maximal value, to find minimal and maximal
in one go, and to calculate standard deviation.
}
\examples{
rm(list=ls())
## min 512 = 2^9 elements
2^21
dat0   <- gpda::rlba(2^21, nthread=64); str(dat0)
result <- gpda::min(dat0$RT); result
result <- gpda::max(dat0$RT); result
result <- gpda::minmax(dat0$RT); result
result <- gpda::sum(dat0$RT); result
result <- gpda::sd(dat0$RT); result

stats::sd(dat0$RT);
base::min(dat0$RT); base::max(dat0$RT)
}
