% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpda.R
\docType{package}
\name{gpda}
\alias{gpda}
\alias{gpda-package}
\title{Probability Density Approximation using GPU}
\description{
Use GPU to calculate probability densities via Monte Carlo simulations.
}
\references{
Holmes, W. (2015). A practical guide to the Probability Density
Approximation (PDA) with improved implementation and error characterization.
\emph{Journal of Mathematical Psychology}, \bold{68-69}, 13--24,
doi: http://dx.doi.org/10.1016/j.jmp.2015.08.006.
}
\author{
Yi-Shin Lin <yishin.lin@utas.edu.au> \cr
Andrew Heathcote <andrew.heathcote@utas.edu.au> \cr
William Holmes <william.holmes@vanderbilts.edu>
}
\keyword{gpda}
