% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{n1PDF}
\alias{n1PDF}
\title{Approximate Node 1 Likelihood in a Two-accumulator LBA Model}
\usage{
n1PDF(x, nsim = 1024, b = 1, A = 0.5, mean_v = c(2.4, 1.6),
  sd_v = c(1, 1), t0 = 0.5, nthread = 64, debug = FALSE)
}
\arguments{
\item{x}{a numeric vector for estimating likelihood.}

\item{nsim}{number of simulations. This must be a power of two.}

\item{b}{threshold. Default is 1.}

\item{A}{starting point upper bound}

\item{mean_v}{mean drift rate. This must be a two-element vector.}

\item{sd_v}{standard deviation of drift rate. This must be a two-element 
vector.}

\item{t0}{non-decision time.}

\item{nthread}{number of threads launched per block. Default is 32.}

\item{debug}{a debugging switch.
Maximum is 1024 in K80 or 512 in other GPU cards.}
}
\value{
a likelihood vector.
}
\description{
This is the density function for the basic 2-accumualtor LBA 
model, sampling drift rates from truncated normal distributions. The 
function approximates model likelihood, instead of using the analytic 
function.
}
\examples{
#################
## n1PDF examples 
#################
data <- seq(0, 3, length.out = 1e3);

## Default parameter set is b = 1, A = 0.5, mean_v = c(2.4, 1.6),
## sd_v = c(1, 1), t0 = 0.5 with nthread = 64 and conduct only 1024 
## simulations
den1 <- gpda::n1PDF(data)
## den2 <- rtdists::n1PDF(data, b=1, A=.5, mean_v=c(2.4, 1.6), sd_v=c(1, 1), 
##                        t0=.5, silent=T)

## Verify that we are not checking near 0 densities
plot(data, den1, type="l")
## lines(data, den2, lwd=2)
## all.equal(den1, den2)
## [1] "Mean relative difference: 0.1675376"

## The approximation is not good enough, so we raise the simulation from
## 2^10 to 2^20 
den1 <- gpda::n1PDF(data, nsim=2^20)
plot(data, den1, type="l")
## lines(data, den2, lwd=2)
## all.equal(den1, den2)
## "Mean relative difference: 0.007108101"
## Now the difference goes down to 2 decimal places below 0.

##########################################
## The cases that rlba does not match dlba 
##########################################
## When approximated PDF is almost perfect with 2^20 simulations,
## one can still observe noise in Bayesian computation. One possible reason
## is the following:   
den1 <- gpda::n1PDF(data, b=.09, A=.07, mean_v=c(-7.37, -4.36), 
sd_v=c(1, 1), t0=.94, nsim=2^20)
## den2 <- rtdists::n1PDF(data, b=.09, A=.07, mean_v=c(-7.37, -4.36), 
## sd_v=c(1, 1), t0=.94, silent=T)
par(mfrow=c(1,2))
plot(data, den1, type="l")
## lines(data, den2, lwd=2)

## plot(data, den2, type="l")
## lines(data, den1, lwd=2)
## all.equal(den1, den2)
## [1] "Mean relative difference: 0.9991495"

## Note both shapes are similar, but dlba method estimates smaller values, 
## relative to rlba method. This happens in behaviourally less plausible  
## parameter sets. It takes longer iterations to smooth out the 
## noise.  That is when a sampler no longer propose these less plausible
## parameter sets. 
}
\references{
Holmes, W. (2015). A practical guide to the Probability Density
Approximation (PDA) with improved implementation and error characterization.
\emph{Journal of Mathematical Psychology}, 68-69, 13--24,
doi: http://dx.doi.org/10.1016/j.jmp.2015.08.006.
}
