% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rlba}
\alias{rlba}
\title{The Random Number Generator of Cannoical Linear Ballistic Accumulator Model}
\usage{
rlba(n, b = 1, A = 0.5, mean_v = c(2.4, 1.6), sd_v = c(1, 1),
  t0 = 0.5, nthread = 32)
}
\arguments{
\item{n}{number of observations. This accepts only integer}

\item{b}{upper bound. (b-A/2) is response caution.}

\item{A}{starting point interval or evidence in accumulator before
beginning of decision process. Start point varies from trial to trial in
the interval [0, A]. Average amount of evidence before evidence
accumulation across trials is A/2.}

\item{mean_v}{drift rate means}

\item{sd_v}{drift rate standard deviations}

\item{t0}{non-decision times}

\item{nThreads}{number of threads launched per block.}
}
\value{
a data frame
}
\description{
This function generates two-accumulator LBA random numbers using GPU.
}
\examples{

}
