% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rlba}
\alias{rlba}
\alias{rlba_n1}
\title{The Random Number Generator of the Basic Linear Ballistic Accumulation Model}
\usage{
rlba(n, b = 1, A = 0.5, mean_v = c(2.4, 1.6), sd_v = c(1, 1),
  t0 = 0.5, nthread = 32, dp = FALSE)

rlba_n1(n, b = 1, A = 0.5, mean_v = c(2.4, 1.6), sd_v = c(1, 1),
  t0 = 0.5, nthread = 32, dp = FALSE)
}
\arguments{
\item{n}{numbers of observation. This accepts only integer scalar.}

\item{b}{threshold.  (b-A/2) is response caution.}

\item{A}{starting point upper bound. This is the upper boundary for an 
uniform distribution that draws a realisation of starting point evidence. 
Presumably, across-trial average starting evidence is A/2.}

\item{mean_v}{means of drift rate. This must be a vector of 2 values.}

\item{sd_v}{drift rate standard deviations. This must be a vector of 2 values.}

\item{t0}{non-decision times. Must be a scalar.}

\item{dp}{whether calculate using double precision. Default is FALSE.}

\item{nthreads}{number of threads launched per block.}
}
\value{
a data frame with first column named RT and second column named R.
}
\description{
This function generates two-accumulator LBA random numbers using GPU.
}
\details{
rlba_n1 draws only node 1 random numbers.
}
\examples{
rm(list=ls())
n <- 2^20
dat1 <- gpda::rlba(n, nthread=64); 
dat2 <- cpda::rlba_test(n)
dat3 <- rtdists::rLBA(n, b=1, A=.5, mean_v=c(2.4, 1.6), sd_v=c(1, 1), t0=.5,
                      silent=TRUE)
names(dat3) <- c("RT","R")

## Trim ----
sum(dat1$RT>5); sum(dat2$RT>5); sum(dat3$RT>5)

dat1 <- dat1[dat1$RT < 5, ]
dat2 <- dat2[dat2$RT < 5, ]
dat3 <- dat3[dat3$RT < 5, ]

dat1c <- dat1[dat1[,2]==1, 1]
dat1e <- dat1[dat1[,2]==2, 1]
dat2c <- dat2[dat2[,2]==1, 1]
dat2e <- dat2[dat2[,2]==2, 1]
dat3c <- dat3[dat3[,2]==1, 1]
dat3e <- dat3[dat3[,2]==2, 1]

den1c <- density(dat1c)
den2c <- density(dat2c)
den3c <- density(dat3c)
den1e <- density(dat1e)
den2e <- density(dat2e)
den3e <- density(dat3e)

## Identical PDFs
par(mfrow=c(1,3))
hist(dat1c, breaks="fd",  col="grey", freq=FALSE, xlab="RT (s)", 
  main="GPU-Choice 1", xlim=c(0, 3)) ## gpu float
lines(den2c, col="red",  lty="dashed",  lwd=1.5) ## cpu
lines(den3c, col="blue", lty="dashed",  lwd=3.0) ## rtdists

hist(dat2c, breaks="fd",  col="grey", freq=FALSE, xlab="RT (s)", 
  main="CPU-Choice 1", xlim=c(0, 3)) ## cpu 
lines(den1c, col="red",  lty="dashed",  lwd=1.5) ## gpu float
lines(den3c, col="blue", lty="dashed",  lwd=3.0) ## rtdists

hist(dat3c, breaks="fd",  col="grey", freq=FALSE, xlab="RT (s)", 
  main="R-Choice 1", xlim=c(0, 3)) ## rtdists
lines(den1c, col="red",  lty="dashed",  lwd=1.5) ## gpu float
lines(den2c, col="blue", lty="dashed",  lwd=3.0) ## cpu
par(mfrow=c(1,1))

plot(den1c$x, den1c$y, type="l")
lines(den1e$x, den1e$y)

lines(den2c$x, den2c$y, col="red", lwd=2, lty="dotted")
lines(den2e$x, den2e$y, col="red", lwd=2, lty="dotted")

lines(den3c$x, den3c$y, col="blue", lwd=2, lty="dashed")
lines(den3e$x, den3e$y, col="blue", lwd=2, lty="dashed")


## Because R script takes a while to run, so I repeated 10 times only
## microbenchmark can still give reliable and precise estimation.
library(microbenchmark)
res <- microbenchmark(gpda::rlba(n, dp=F),
                      gpda::rlba(n, dp=T),
                      cpda::rlba_test(n),
                      rtdists::rLBA(n, b=1, A=.5, mean_v=c(2.4, 1.6),
sd_v=c(1, 1), t0=.5, silent=TRUE), times=10L)

## Unit: milliseconds
##                  expr    min       lq     mean   median       uq      max neval cld
## gpda::rlba(n, dp=F)     8.31     8.47     9.17     8.53     9.21    11.59    10   a
## gpda::rlba(n, dp=T)    11.86    11.96    12.31    12.06    12.17    14.86    10   a
## cpda::rlba_test(n)    201.83   202.05   209.32   202.84   225.03   225.72    10   b
## rtdists::rLBA(n, .) 13521.67 13614.74 13799.59 13770.78 13919.77 14177.51    10   c

rm(list=ls())
n <- 2^20; n
dat1 <- gpda::rlba_n1(n, nthread=64, dp=T); str(dat1)
dat2 <- gpda::rlba_n1(n, nthread=64, dp=F); str(dat2)

res <- microbenchmark(
gpda::rlba_n1(n, dp=F),
gpda::rlba_n1(n, dp=T), times=10L)

res
## Unit: milliseconds
##                     expr       min        lq     mean   median       uq
## gpda::rlba_n1(n, dp = F)  9.572601  9.949197 17.33809 10.82227 11.41039
## gpda::rlba_n1(n, dp = T) 13.774382 14.742762 21.77835 15.07504 15.45775
##      max neval cld
## 44.93433    10   a
## 49.63786    10   a

}
