% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sum_gpu}
\alias{sum_gpu}
\title{GPU Math Functions}
\usage{
sum_gpu(x, debug = TRUE)
}
\arguments{
\item{x}{a data vector.}

\item{debug}{whether to print debug information}
}
\value{
a scalar or a pair of scalar values.
}
\description{
sum, min, max, minmax and sd are 5 basic mathematical functions to sum a
vector, to find minimal or maximal value, to find minimal and maximal
in one go, and to calculate standard deviation.
}
\examples{
rm(list=ls())
## min 512 = 2^9 elements

dat0   <- gpda::rlba(2^21, nthread=64); str(dat0)
result <- gpda::min(dat0$RT); result
result <- gpda::max(dat0$RT); result
result <- gpda::minmax(dat0$RT); result
result <- gpda::sum(dat0$RT); result
result <- gpda::sd(dat0$RT); result

stats::sd(dat0$RT);
base::min(dat0$RT); base::max(dat0$RT)
}
